nats

+++ protocol/location.go
@@ -1,7 +1,7 @@
 package protocol
 
 type Location struct {
-	Longitude float32 `json:"longitude"` // exp: 54.306
-	Latitude  float32 `json:"latitude"`  // exp: 10.149
-	Heading   int     `json:"heading"`   // in degree
+	Longitude int `json:"longitude"` // exp: 54.306 * 3600000 = longitude
+	Latitude  int `json:"latitude"`  // exp: 10.149 * 3600000 = latitude
+	Heading   int `json:"heading"`   // in degree
 }
